pyproject.toml
[project]
name = "sqlite-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.4",
    "typer>=0.12.3",
]
